Do not break local start order of application
Before it was impossible to have any starting order
of applications, because swarm would always tries to
start remote swarm, before any application is started.

For example, if your distributed via swarm process X depends
on application Y, than you want to start application, like this:
Y, swarm. If remote swarm is quicklier, than local swarm app is
started before application Y, which starts distribution and handoff
of processes to a node, which isn't ready to process this processes,
because applications on the node are not started yet.