Ensure sync occurs upon initial cluster join
When joining a cluster after initial startup, the tracker will be in
tracking state, and needs an opportunity to sync with the cluster when
it joins. Previously, this synchronization only happened during startup
or during anti-entropy passes, but this commit ensures that if this
initial join occurs during tracking, that it is caught and handled like
the cluster_wait -> cluster_join transition, ensuring a sync right away.

See #62