Address synchronization issues
Problem: If multiple nodes are started, and do not form a cluster prior
to the initial sync_nodes_timeout, when they later form a cluster, they
are not synchronized properly and it could take quite some time for them
to sync. In addition, the existing synchronization logic was flawed in
that it was not well coordinated and did not ensure that both sides of
the sync received an up-to-date registry.

The following changes address the issues above:

- Rather than rolling a die to choose which node would begin a sync, and
coordinating on that, we instead include the tracker's clock in the sync
request, and based on the clock, the node which should begin the sync
process is chosen. This is deterministic and requires no extra
coordination, allowing us to remove the "tiebreaker" code which existed.
- When the first sync is completed, the acknowledgement to the
originating node includes the updated registry snapshot from the node
which is now fully up to date, allowing the originating node to then
sync itself with that up to date registry, ensuring both sides are
completely synchronized when the sync completes.
- Since the coordination is deterministic, we avoid the issue with
spurious sync requests or acknowledgements from arriving after the
tracker has moved out of the syncing state.

See #62 for more information