Fix replication event for removing metadata
It seems like the removal of metadata in the Swarm.leave call
is not properly replicated.

The handle_replica_event for :update_meta is always merging
the local with the remote meta data which means existing
keys always stay in the meta dictionary. Thus, the removal
of keys is never taking affect on remote nodes.

Broadcasting specific add_meta and remove_meta replication
events instead of a more general update_meta event which gives
remote nodes more context to apply the correct state.

Updated also adding metadata, even though it was working, but
it might be better to keep these in-sync.